canal/canal/urls.py
"""canal URL Configuration"""
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from django.views.generic import RedirectView

urlpatterns = [
    path('', RedirectView.as_view(url='/chat/', permanent=False)),
    path('admin/', admin.site.urls),
    path('chat/', include(('chat.urls', 'chat'), namespace='chat')),
] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
