requirements/development.txt
-r live.txt
channels_redis
django-nose
django-webtest
ipdb
ipython
nose
spec
webtest
werkzeug